ThreadedGenerator: tweak threading algorithm
to have a fixed 30% chance of making a message be a follow-up to the
previously generated message.  This provides some burstiness to threads.

One particular result is to remove the unfortunate artifact where the
first approximately $nthreads messages in a generated sequence are all
from different threads, so there's little difference between the
exemplars that xconvsort finds in conversation mode or message mode.