Speed up tests which deliberately kill master
Some tests will deliberately encode impossible things in cyrus.conf and
expect master to exit early before all the configured services are
available.  Currently this causes a long timeout in
Instance::_start_master() because it waits for all the configured
services to appear in the netstat output.  Add an Instance::is_running()
check there to break the loop early.